made chat id 64 bits

+++ types.go
@@ -54,7 +54,7 @@ type TelegramUpdateResult struct {
 }
 
 type TelegramOutgoingMessage struct {
-	ChatID int    `json:"chat_id"`
+	ChatID int64  `json:"chat_id"`
 	Text   string `json:"text"`
 }
